removed streaming module in favor of npstreams
