+++ requirements.txt
@@ -2,5 +2,6 @@ numpy>= 1.11.2
 scipy>= 0.17
 scikit-image
 pywavelets>=0.5.2
+npstreams>= 0.1.2
 spglib
 pycifrw
