+++ skued/image/__init__.py
@@ -2,7 +2,6 @@
 """ Diffraction image analysis """
 
 from .powder import angular_average, powder_center
-from .alignment import align, shift_image, diff_register
+from .alignment import align, shift_image, diff_register, ialign
 from .symmetry import nfold_symmetry, nfold
 from .correlation import mnxc2
-from .streaming import ialign, iaverage, isem, istd, ivar
